User Index Cleanup

+++ resources/views/users/table.blade.php
@@ -3,30 +3,27 @@
         <thead>
         <tr>
             <th>Name</th>
-        <th>Role Id</th>
-        <th>Email</th>
-        <th>Email Verified At</th>
-        <th>Password</th>
-        <th>Remember Token</th>
-            <th colspan="3">Action</th>
+            <th>Email</th>
+            <th>User Level</th>
+            @if(Auth::user()->role_id < 3)
+                <th colspan="3">Action</th>
+            @endif
         </tr>
         </thead>
         <tbody>
         @foreach($users as $user)
             <tr>
-                <td>{{ $user->name }}</td>
-            <td>{{ $user->role_id }}</td>
+                <td>
+                    <a href="{{ route('users.show', [$user->id]) }}">
+                        {{ $user->name }}
+                    </a>
+                </td>
             <td>{{ $user->email }}</td>
-            <td>{{ $user->email_verified_at }}</td>
-            <td>{{ $user->password }}</td>
-            <td>{{ $user->remember_token }}</td>
+            <td>{{ $user->role['name'] }}</td>
+            @if(Auth::user()->role_id < 3)
                 <td width="120">
                     {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
                     <div class='btn-group'>
-                        <a href="{{ route('users.show', [$user->id]) }}"
-                           class='btn btn-default btn-xs'>
-                            <i class="far fa-eye"></i>
-                        </a>
                         <a href="{{ route('users.edit', [$user->id]) }}"
                            class='btn btn-default btn-xs'>
                             <i class="far fa-edit"></i>
@@ -35,6 +32,7 @@
                     </div>
                     {!! Form::close() !!}
                 </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
